requirements.txt
Flask==2.3.3
supabase==1.0.3
sentence-transformers==2.2.2
python-dotenv==1.0.0
Flask-CORS==4.0.0
bcrypt==4.0.1
requests==2.31.0
numpy==1.24.3
scikit-learn==1.3.0
